Refactor roamer systems and add time scaling
Replaces RoamerMovementSystem with RoamerBehaviourSystem and StriderBehaviourSystem, and introduces RoamerKillSystem for handling roamer kills. Adds time scaling for debug builds, affecting deltaTime and related systems. Refactors restart menu and settings menu to use tag components and toggling logic. Updates AnomalyRoamerComponent to support target selection. Minor code cleanups and improvements throughout.

+++ src/assemblers/scenes/comms_scene/radar_object.cpp
@@ -28,7 +28,7 @@
 void Construct::RadarObject(entt::registry& registry, ResourceStore& resourceStore)
 {
 	constexpr float ATTRACTION_REDUCTION = 1.0f;
-	constexpr float POWER_USAGE = 500.0f;
+	constexpr uint16_t POWER_USAGE = 500u;
 
 	const entt::entity entity = registry.create();
